+++ packages/article-image/__tests__/mocks.native.js
@@ -1,0 +1,3 @@
+jest.mock("@times-components/image", () => ({
+  ModalImage: "ModalImage"
+}));