+++ packages/article-main-standard/article-main-standard.showcase.js
@@ -10,6 +10,7 @@ import {
 } from "@times-components/provider-test-tools";
 import { sections } from "@times-components/storybook";
 import { scales, themeFactory } from "@times-components/styleguide";
+import Responsive from "@times-components/responsive";
 import storybookReporter from "@times-components/tealium-utils";
 import Article from "./src/article-main-standard";
 
@@ -36,46 +37,52 @@ const renderArticle = ({
   scale,
   section
 }) => (
-  <ArticleProvider debounceTimeMs={0} id={id}>
-    {({ article, isLoading, error, refetch }) => (
-      <Context.Provider
-        value={{
-          makeArticleUrl,
-          theme: {
-            ...themeFactory(section, templateName),
-            scale: scale || defaults.theme.scale
-          }
-        }}
-      >
-        <Article
-          adConfig={adConfig}
-          analyticsStream={analyticsStream}
-          article={article}
-          error={error}
-          isLoading={isLoading}
-          onAuthorPress={preventDefaultedAction(decorateAction)(
-            "onAuthorPress"
-          )}
-          onCommentGuidelinesPress={preventDefaultedAction(decorateAction)(
-            "onCommentGuidelinesPress"
-          )}
-          onCommentsPress={preventDefaultedAction(decorateAction)(
-            "onCommentsPress"
-          )}
-          onLinkPress={preventDefaultedAction(decorateAction)("onLinkPress")}
-          onRelatedArticlePress={preventDefaultedAction(decorateAction)(
-            "onRelatedArticlePress"
-          )}
-          onTopicPress={preventDefaultedAction(decorateAction)("onTopicPress")}
-          onTwitterLinkPress={preventDefaultedAction(decorateAction)(
-            "onTwitterLinkPress"
-          )}
-          onVideoPress={preventDefaultedAction(decorateAction)("onVideoPress")}
-          refetch={refetch}
-        />
-      </Context.Provider>
-    )}
-  </ArticleProvider>
+  <Responsive>
+    <ArticleProvider debounceTimeMs={0} id={id}>
+      {({ article, isLoading, error, refetch }) => (
+        <Context.Provider
+          value={{
+            makeArticleUrl,
+            theme: {
+              ...themeFactory(section, templateName),
+              scale: scale || defaults.theme.scale
+            }
+          }}
+        >
+          <Article
+            adConfig={adConfig}
+            analyticsStream={analyticsStream}
+            article={article}
+            error={error}
+            isLoading={isLoading}
+            onAuthorPress={preventDefaultedAction(decorateAction)(
+              "onAuthorPress"
+            )}
+            onCommentGuidelinesPress={preventDefaultedAction(decorateAction)(
+              "onCommentGuidelinesPress"
+            )}
+            onCommentsPress={preventDefaultedAction(decorateAction)(
+              "onCommentsPress"
+            )}
+            onLinkPress={preventDefaultedAction(decorateAction)("onLinkPress")}
+            onRelatedArticlePress={preventDefaultedAction(decorateAction)(
+              "onRelatedArticlePress"
+            )}
+            onTopicPress={preventDefaultedAction(decorateAction)(
+              "onTopicPress"
+            )}
+            onTwitterLinkPress={preventDefaultedAction(decorateAction)(
+              "onTwitterLinkPress"
+            )}
+            onVideoPress={preventDefaultedAction(decorateAction)(
+              "onVideoPress"
+            )}
+            refetch={refetch}
+          />
+        </Context.Provider>
+      )}
+    </ArticleProvider>
+  </Responsive>
 );
 
 const mockArticle = ({