+++ packages/edition-slices/__tests__/shared.native.js
@@ -1,0 +1,28 @@
+import {
+  addSerializers,
+  compose,
+  enzymeRenderedSerializer,
+  minimaliseTransform,
+  minimalNativeTransform,
+  print
+} from "@times-components/jest-serializer";
+import shared from "./shared.base";
+
+jest.mock("@times-components/image", () => "Image");
+jest.mock("@times-components/gradient", () => "Gradient");
+
+export default () => {
+  addSerializers(
+    expect,
+    enzymeRenderedSerializer(),
+    compose(
+      print,
+      minimalNativeTransform,
+      minimaliseTransform(
+        (value, key) => key === "style" || key.includes("Class")
+      )
+    )
+  );
+
+  shared();
+};