+++ android-app/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.1](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.11.0...@times-components/android-app@0.11.1) (2019-01-23)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+# [0.11.0](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.10.84...@times-components/android-app@0.11.0) (2019-01-22)
+
+
+### Features
+
+* key facts tablet layout ([#1612](https://github.com/newsuk/times-components/issues/1612)) ([0793a0a](https://github.com/newsuk/times-components/commit/0793a0a))
+
+
+
+
+
+## [0.10.84](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.10.83...@times-components/android-app@0.10.84) (2019-01-22)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
 ## [0.10.83](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.10.82...@times-components/android-app@0.10.83) (2019-01-21)
 
 **Note:** Version bump only for package @times-components/android-app